Search Populating Tables
Had to change the code for search to populate each table cell. Also changed some of the css and added some css classes for the contacts data/buttons. Also commented out a piece of code in contact.html that was leaving a gray box at the top of the page.